docshell/test/mochitest/test_bug680257.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=680257
-->
<head>
  <title>Test for Bug 680257</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=680257">Mozilla Bug 680257</a>

<script type="application/javascript">

SimpleTest.waitForExplicitFinish();

var popup = window.open("file_bug680257.html");

var gTestContinuation = null;
function continueAsync() {
  setTimeout(function() { gTestContinuation.next(); })
}

// The popup will call into popupLoaded() once it loads.
function popupLoaded() {
  // runTests() needs to be called from outside popupLoaded's onload handler.
  // Otherwise, the navigations we do in runTests won't create new SHEntries.
  SimpleTest.executeSoon(function() {
    if (!gTestContinuation) {
      gTestContinuation = runTests();
    }
    gTestContinuation.next();
  });
}

function* runTests() {
  checkPopupLinkStyle(false, "Initial");

  popup.location.hash = "a";
  checkPopupLinkStyle(true, "After setting hash");

  popup.history.back();
  continueAsync();
  yield;

  checkPopupLinkStyle(false, "After going back");

  popup.history.forward();
  continueAsync();
  yield;
  checkPopupLinkStyle(true, "After going forward");

  popup.close();
  SimpleTest.finish();
}

function checkPopupLinkStyle(isTarget, desc) {
  var link = popup.document.getElementById("a");
  var style = popup.getComputedStyle(link);
  var color = style.getPropertyValue("color");

  // Color is red if isTarget, black otherwise.
  if (isTarget) {
    is(color, "rgb(255, 0, 0)", desc);
  } else {
    is(color, "rgb(0, 0, 0)", desc);
  }
}

</script>
</body>
</html>
